+++ plugin_tooling/src-test/dtool/engine/operations/CompletionOperation_Test.java
@@ -184,22 +184,22 @@ public class CompletionOperation_Test extends CommonDToolOperation_Test {
 		offset = MODULE_Contents.indexOf("/*N*/");
 		testCodeCompletionProposals(MODULE_FilePath, offset,
 			new LangCompletionResult(list(
-				proposal(offset, 0, "", "void", 
-					CompletionProposalKind.NATIVE, att(), null, "")
+				proposal(offset-4, 4, "void", "void", 
+					CompletionProposalKind.NATIVE, att(), null, "void")
 			))
 		);
 		
 		offset = MODULE_Contents.indexOf("/*CC_1*/");
 		testCodeCompletionProposals(MODULE_FilePath, offset,
 			new LangCompletionResult(list(
-				proposal(offset, 0, "", "foo() : void", 
-					CompletionProposalKind.FUNCTION, att(), "_dummy2", "()"),
-				proposal(offset, 0, "", "foo(int a) : void", 
-					CompletionProposalKind.FUNCTION, att(), "_dummy2", "(a)", sr(1, 1)),
-				proposal(offset, 0, "", "foo(int a, string str) : void", 
-					CompletionProposalKind.FUNCTION, att(), "_dummy2", "(a, str)", sr(1, 1), sr(4, 3)),
-				proposal(offset, 0, "TemplateFn", "fooTemplateFn(T) (T param) : int", 
-					CompletionProposalKind.FUNCTION, att(), "_dummy2", "TemplateFn(param)", sr(11, 5))
+				proposal(offset-3, 3, "foo", "foo() : void", 
+					CompletionProposalKind.FUNCTION, att(), "_dummy2", "foo()"),
+				proposal(offset-3, 3, "foo", "foo(int a) : void", 
+					CompletionProposalKind.FUNCTION, att(), "_dummy2", "foo(a)", sr(4, 1)),
+				proposal(offset-3, 3, "foo", "foo(int a, string str) : void", 
+					CompletionProposalKind.FUNCTION, att(), "_dummy2", "foo(a, str)", sr(4, 1), sr(7, 3)),
+				proposal(offset-3, 3, "fooTemplateFn", "fooTemplateFn(T) (T param) : int", 
+					CompletionProposalKind.FUNCTION, att(), "_dummy2", "fooTemplateFn(param)", sr(14, 5))
 			))
 		);