+++ documentation/ChangeLog.md
@@ -3,6 +3,7 @@
 ### (NextVersion)
  * Fixed: hover for problem errors not showing if error is over a string literal source.
  * Fixed #112: build errors no longer removed when editing a source file.
+ * Fixed #113: Content Assist proposal has incorrect casing
  
 ### DDT 0.12.0